Build test tasks at build time instead of CodeTaskFactory
