+++ src/Build.UnitTests/BackEnd/TaskBuilder_Tests.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Reflection;
 using System.Threading.Tasks;
 using System.Xml;
 using Microsoft.Build.BackEnd;
@@ -30,121 +31,6 @@ namespace Microsoft.Build.UnitTests.BackEnd
     /// </summary>
     public class TaskBuilder_Tests : ITargetBuilderCallback
     {
-
-#if FEATURE_CODEDOM
-        /// <summary>
-        /// Task definition for a task that outputs items containing null metadata.
-        /// </summary>
-        private static string s_nullMetadataTaskContents =
-@"using Microsoft.Build.Framework;
-using Microsoft.Build.Utilities;
-using System.Collections;
-using System.Collections.Generic;
-
-namespace NullMetadataTask
-{
-    public class NullMetadataTask : Task
-    {
-        [Output]
-        public ITaskItem[] OutputItems
-        {
-            get;
-            set;
-        }
-
-        public override bool Execute()
-        {
-            OutputItems = new ITaskItem[1];
-
-            IDictionary<string, string> metadata = new Dictionary<string, string>();
-            metadata.Add(""a"", null);
-
-            OutputItems[0] = new TaskItem(""foo"", (IDictionary)metadata);
-
-            return true;
-        }
-    }
-}
-";
-
-        /// <summary>
-        /// Task definition for task that outputs items in a variety of ways, used to 
-        /// test definition of the DefiningProject* metadata for task outputs. 
-        /// </summary>
-        private static string s_itemCreationTaskContents =
-            @"using Microsoft.Build.Framework;
-using Microsoft.Build.Utilities;
-using System.Collections;
-using System.Collections.Generic;
-
-namespace ItemCreationTask
-{
-    public class ItemCreationTask : Task
-    {
-        public ITaskItem[] InputItemsToPassThrough
-        {
-            get;
-            set;
-        }
-
-        public ITaskItem[] InputItemsToCopy
-        {
-            get;
-            set;
-        }
-
-        [Output]
-        public ITaskItem[] PassedThroughOutputItems
-        {
-            get;
-            set;
-        }
-
-        [Output]
-        public ITaskItem[] CreatedOutputItems
-        {
-            get;
-            set;
-        }
-
-        [Output]
-        public ITaskItem[] CopiedOutputItems
-        {
-            get;
-            set;
-        }
-
-        [Output]
-        public string OutputString
-        {
-            get;
-            set;
-        }
-
-        public override bool Execute()
-        {
-            PassedThroughOutputItems = InputItemsToPassThrough;
-
-            CopiedOutputItems = new ITaskItem[InputItemsToCopy.Length];
-
-            for (int i = 0; i < InputItemsToCopy.Length; i++)
-            {
-                CopiedOutputItems[i] = new TaskItem(InputItemsToCopy[i]);
-            }
-
-            CreatedOutputItems = new ITaskItem[2];
-            CreatedOutputItems[0] = new TaskItem(""Foo"");
-            CreatedOutputItems[1] = new TaskItem(""Bar"");
-
-            OutputString = ""abc;def;ghi"";
-
-            return true;
-        }
-    }
-}
-";
-#endif
-
         /// <summary>
         /// The mock component host and logger
         /// </summary>
@@ -588,7 +474,7 @@ namespace ItemCreationTask
         [Trait("Category", "mono-osx-failing")]
         public void NullMetadataOnOutputItems()
         {
-            string customTaskPath = CustomTaskHelper.GetAssemblyForTask(s_nullMetadataTaskContents);
+            string customTaskPath = Assembly.GetExecutingAssembly().Location;
 
             string projectContents = @"<Project ToolsVersion=`msbuilddefaulttoolsversion` xmlns=`msbuildnamespace`>
   <UsingTask TaskName=`NullMetadataTask` AssemblyFile=`" + customTaskPath + @"` />
@@ -613,17 +499,7 @@ namespace ItemCreationTask
         [Trait("Category", "mono-osx-failing")]
         public void NullMetadataOnLegacyOutputItems()
         {
-            string referenceAssembliesPath = ToolLocationHelper.GetPathToDotNetFrameworkReferenceAssemblies(TargetDotNetFrameworkVersion.VersionLatest);
-
-            if (String.IsNullOrEmpty(referenceAssembliesPath))
-            {
-                // fall back to the .NET Framework -- they should always exist there. 
-                referenceAssembliesPath = ToolLocationHelper.GetPathToDotNetFramework(TargetDotNetFrameworkVersion.VersionLatest);
-            }
-
-            string[] referenceAssemblies = new string[] { Path.Combine(referenceAssembliesPath, "System.dll"), Path.Combine(referenceAssembliesPath, "Microsoft.Build.Framework.dll"), Path.Combine(referenceAssembliesPath, "Microsoft.Build.Utilities.v4.0.dll") };
-
-            string customTaskPath = CustomTaskHelper.GetAssemblyForTask(s_nullMetadataTaskContents, referenceAssemblies);
+            string customTaskPath = Assembly.GetExecutingAssembly().Location;
 
             string projectContents = @"<Project ToolsVersion=`msbuilddefaulttoolsversion` xmlns=`msbuildnamespace`>
   <UsingTask TaskName=`NullMetadataTask` AssemblyFile=`" + customTaskPath + @"` />
@@ -732,7 +608,7 @@ namespace ItemCreationTask
         [Fact]
         public void ValidateDefiningProjectMetadataOnTaskOutputs()
         {
-            string customTaskPath = CustomTaskHelper.GetAssemblyForTask(s_itemCreationTaskContents);
+            string customTaskPath = Assembly.GetExecutingAssembly().Location;
             ValidateDefiningProjectMetadataOnTaskOutputsHelper(customTaskPath);
         }
 
@@ -745,16 +621,7 @@ namespace ItemCreationTask
         [Trait("Category", "mono-osx-failing")]
         public void ValidateDefiningProjectMetadataOnTaskOutputs_LegacyItems()
         {
-            string referenceAssembliesPath = ToolLocationHelper.GetPathToDotNetFrameworkReferenceAssemblies(TargetDotNetFrameworkVersion.VersionLatest);
-
-            if (String.IsNullOrEmpty(referenceAssembliesPath))
-            {
-                referenceAssembliesPath = ToolLocationHelper.GetPathToDotNetFramework(TargetDotNetFrameworkVersion.VersionLatest);
-            }
-
-            string[] referenceAssemblies = new string[] { Path.Combine(referenceAssembliesPath, "System.dll"), Path.Combine(referenceAssembliesPath, "Microsoft.Build.Framework.dll"), Path.Combine(referenceAssembliesPath, "Microsoft.Build.Utilities.v4.0.dll") };
-
-            string customTaskPath = CustomTaskHelper.GetAssemblyForTask(s_itemCreationTaskContents, referenceAssemblies);
+            string customTaskPath = Assembly.GetExecutingAssembly().Location;
             ValidateDefiningProjectMetadataOnTaskOutputsHelper(customTaskPath);
         }
